requirements.txt
asgiref==3.9.1
channels==4.1.0  # Понижена версия для совместимости с Django 5.2
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
daphne==4.1.0  # Фиксированная версия
drf-spectacular==0.27.1  # Фиксированная версия
django-cors-headers==4.3.1  # Явное указание пакета
psycopg2-binary==2.9.9  # Стабильная версия
websockets==12.0
Pillow==10.3.0
dotenv
celery